OpaqueSQLSuiteBase
